test(index-controller): remove unnecessary service stub
