+++ app/components/optimized-power-select/trigger/template.hbs
@@ -1,30 +1,36 @@
-{{#if @select.selected}}
-  {{#if @selectedItemComponent}}
-    {{component
-      @selectedItemComponent
-      option=(readonly @select.selected)
-      select=(readonly @select)
-    }}
+<div class={{@triggerClass}}>
+  {{#if @select.selected}}
+    {{#if @selectedItemComponent}}
+      {{component
+        @selectedItemComponent
+        option=(readonly @select.selected)
+        select=(readonly @select)
+      }}
+    {{else}}
+      <span class="ember-power-select-selected-item">
+        {{#if @extra.selectedTemplate}}
+          {{component
+            (ensure-safe-component @extra.selectedTemplate)
+            selected=@select.selected
+          }}
+        {{else}}
+          {{yield @select.selected @select}}
+        {{/if}}
+      </span>
+    {{/if}}
+    {{#if (and @allowClear (not @select.disabled))}}
+      <span
+        role="button"
+        class="ember-power-select-clear-btn"
+        {{on "mouseup" this.clear}}
+        {{on "ontouchend" this.clear}}
+      >&times;</span>
+    {{/if}}
   {{else}}
-    <span class="ember-power-select-selected-item">
-      {{#if @extra.selectedTemplate}}
-        {{component (ensure-safe-component @extra.selectedTemplate) selected=@select.selected}}
-      {{else}}
-        {{yield @select.selected @select}}
-      {{/if}}
-    </span>
+    {{component
+      (ensure-safe-component @placeholderComponent)
+      placeholder=@placeholder
+    }}
   {{/if}}
-  {{#if (and @allowClear (not @select.disabled))}}
-    <span
-      class="ember-power-select-clear-btn"
-      {{on "mousedown" this.clear}}
-      {{on "ontouchstart" this.clear}}
-    >&times;</span>
-  {{/if}}
-{{else}}
-  {{component
-    (ensure-safe-component @placeholderComponent)
-    placeholder=@placeholder
-  }}
-{{/if}}
-<span class="ember-power-select-status-icon"></span>
+  <span class="ember-power-select-status-icon"></span>
+</div>
